% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_filters.R
\name{add_filter_tiddler_to_tiddlywiki}
\alias{add_filter_tiddler_to_tiddlywiki}
\title{Allow certain filters via api}
\usage{
add_filter_tiddler_to_tiddlywiki(allow_filters)
}
\arguments{
\item{allow_filters}{A vector with the filters to allow}
}
\description{
This function adds specific tiddler to the server that later allow you to do filtering
}
