% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_filters.R
\name{add_filters}
\alias{add_filters}
\title{Add Specific Filters to TiddlyWiki Server}
\usage{
add_filters(mws = FALSE)
}
\arguments{
\item{mws}{Logical. If TRUE, sets the TiddlyWiki options to use the "tide" recipe for a MultiWikiServer.. Default is FALSE.}
}
\description{
This internal function adds a specific set of filters to the TiddlyWiki server,
enabling filtering functionality via the API.
}
\keyword{internal}
