% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_journals.R
\name{add_journals}
\alias{add_journals}
\title{Add Daily Journal Tiddlers to TiddlyWiki Server}
\usage{
add_journals(
  start_date = format(Sys.Date(), "\%Y-\%m-\%d"),
  end_date = format(Sys.Date(), "\%Y-\%m-\%d"),
  mws = FALSE
)
}
\arguments{
\item{start_date}{Character. The start date in "YYYY-MM-DD" format. Defaults to today's date.}

\item{end_date}{Character. The end date in "YYYY-MM-DD" format. Defaults to today's date.}

\item{mws}{Logical. If TRUE, sets the TiddlyWiki options to use the "tide-journal" recipe for a MultiWikiServer. Default is FALSE.}
}
\description{
This internal function adds daily journal tiddlers to the TiddlyWiki server for a specified date range.
Each journal entry is represented as a tiddler with the date as the title and tagged as "Journal".
}
\keyword{internal}
