% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_settings.R
\name{add_settings}
\alias{add_settings}
\title{Add Settings to TiddlyWiki Server}
\usage{
add_settings(tidepath, mws = FALSE)
}
\arguments{
\item{tidepath}{Character. The path where the tide folder is located.}

\item{mws}{Logical. If TRUE, sets the TiddlyWiki options to use the "tide" recipe for a MultiWikiServer. Default is FALSE.}
}
\description{
This internal function adds specific settings tiddlers to the TiddlyWiki server,
which configure paths and MultiWikiServer (MWS) settings.
}
\keyword{internal}
