% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create Folder Structure}
\usage{
create_folders(mypath)
}
\arguments{
\item{mypath}{Character. A relative path from the current working directory where the folders should be created.}
}
\description{
This internal function creates a specific folder structure at the specified relative path.
}
\keyword{internal}
