% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mws_server.R
\name{create_mws_server}
\alias{create_mws_server}
\title{Create MultiWikiServer (MWS) for TiddlyWiki}
\usage{
create_mws_server(mws_path, tiddlywiki)
}
\arguments{
\item{mws_path}{Character. The path where the MWS folder should be created.}

\item{tiddlywiki}{Character. The path to the TiddlyWiki installation.}
}
\description{
This internal function sets up a MultiWikiServer (MWS) for TiddlyWiki at the specified path,
initializing necessary configurations and plugins.
}
\keyword{internal}
