% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_run_txt.R
\name{make_run_txt}
\alias{make_run_txt}
\title{Generate Command Text to Start the TiddlyWiki Server}
\usage{
make_run_txt(tidepath = "myTide", mws = FALSE, tiddlywiki = NA, wikipath = NA)
}
\arguments{
\item{tidepath}{Character. The path where the tide folder is located. Default is "myTide".}

\item{mws}{Logical. Indicates whether to set up the TiddlyWiki server in Multi-User Server mode. Default is FALSE.}

\item{tiddlywiki}{Character. The path to the TiddlyWiki installation. Default is NA.}

\item{wikipath}{Character. The path where the wiki data will be stored. Default is NA.}
}
\value{
A character string that contains the command to start the TiddlyWiki server with the specified settings.
}
\description{
This internal function constructs the command text needed to start the TiddlyWiki server
based on the provided parameters. The generated command text can be used to start the server
with specific configurations.
}
\examples{
# Example usage:
make_run_txt()
make_run_txt(tidepath = "customPath", mws = TRUE, tiddlywiki = "path/to/tiddlywiki")
}
\keyword{internal}
