% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tides.R
\name{run_tides}
\alias{run_tides}
\title{Import and run tides}
\usage{
run_tides(force_run = FALSE)
}
\arguments{
\item{force_run}{Also run scripts tagged with \code{DoNotRun}}
}
\description{
This function runs the tides in your tiddlywiki
}
