% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_server.R
\name{start_server}
\alias{start_server}
\title{Start the TiddlyWiki Server}
\usage{
start_server(tidepath = "myTide", mws = FALSE, tiddlywiki = NA, wikipath = NA)
}
\arguments{
\item{tidepath}{Character. The path where the tide folder is located. Default is "myTide".}

\item{mws}{Logical. Indicates whether to set up the TiddlyWiki server in MultiWikiServer mode. Default is FALSE.}

\item{tiddlywiki}{Character. The path to the TiddlyWiki installation. Default is NA.}

\item{wikipath}{Character. The path where the wiki data will be stored. Default is NA.}
}
\value{
An object representing the process running the TiddlyWiki server.
}
\description{
This function starts the TiddlyWiki server with the specified configurations.
It can start the server either in MultiWikiServer (MWS) mode or regular mode.
}
\examples{
# Start the server with default settings
p <- start_server()

# Start the server with custom settings
p <- start_server(tidepath = "customPath", mws = TRUE, tiddlywiki = "path/to/tiddlywiki", wikipath = "path/to/wiki")
}
