% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_server.R
\name{stop_server}
\alias{stop_server}
\title{Stop the TiddlyWiki Server}
\usage{
stop_server(p)
}
\arguments{
\item{p}{Process object. The process object representing the running TiddlyWiki server.}
}
\description{
This function stops the TiddlyWiki server by interrupting the process.
}
\examples{
# Start the server
p <- start_server()

# Stop the server
stop_server(p)
}
