% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_filters.R
\name{add_filter_tiddler_to_tiddlywiki}
\alias{add_filter_tiddler_to_tiddlywiki}
\title{Add Filter Tiddlers to TiddlyWiki Server}
\usage{
add_filter_tiddler_to_tiddlywiki(allow_filters, mws = FALSE)
}
\arguments{
\item{allow_filters}{Character vector. A vector containing the filters to be added.}

\item{mws}{Logical. If TRUE, sets the TiddlyWiki options to use the "tide" recipe for a MultiWikiServer. Default is FALSE.}
}
\description{
This internal function adds specified filter tiddlers to the TiddlyWiki server,
enabling the use of those filters via the API.
}
\keyword{internal}
