% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create Folders}
\usage{
create_folders(mypath)
}
\arguments{
\item{mypath}{A relative path from the current working directory where the folders should be created.}
}
\description{
This function creates a folder structure at the specified relative path.
}
